requirements.txt
uvicorn[standard]
fastapi==0.115.12
passlib==1.7.4
pydantic[email]==2.11.4
PyJWT==2.10.1
python-dotenv==1.1.0
slowapi==0.1.9
SQLAlchemy==2.0.41
starlette==0.46.2
psycopg2-binary
python-multipart
email-validator
PyMuPDF
attrs==25.3.0
Automat==25.4.16
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.31.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==11.2.1
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
Scrapy==2.12.0
sentence-transformers==4.1.0
service-identity==24.2.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tldextract==5.2.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
Twisted==24.11.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
w3lib==2.3.1
zope.interface==7.2
backend==0.2.4.1
fuzzywuzzy